pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meeting-entropy"
version = "1.0.0"
description = "Real-time corporate noise detection for meetings. Because someone had to build this."
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [
    { name = "Julien Mer — kevin-sigmoid-org" },
]
keywords = ["meeting", "entropy", "buzzword", "detection", "whisper", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Office/Business",
]
dependencies = [
    "faster-whisper>=1.0.0",
    "pyaudio>=0.2.14",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
]

[project.scripts]
meeting-entropy = "meeting_entropy.cli:cli"

[project.urls]
Homepage = "https://github.com/kevin-sigmoid-org/meeting-entropy"
Repository = "https://github.com/kevin-sigmoid-org/meeting-entropy"
Issues = "https://github.com/kevin-sigmoid-org/meeting-entropy/issues"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
